Added side room base and decor, fixed wizard door, added new vicente interactables
Yes

+++ No Hesitation Move!/Assets/Scripts/positionRendererSorter.cs
@@ -6,7 +6,7 @@ public class positionRendererSorter : MonoBehaviour {
 	[SerializeField]
 	private int sortingOrderBase = 5000;
 	[SerializeField]
-	private int offset = 0;
+	private float offset = 0f;
 	[SerializeField]
 	private bool runOnlyOnce = false;
